+++ BankMail.lua
@@ -24,7 +24,8 @@ local defaults = {
     enableAutoSwitchOnBank = false,
     enableCoinSubject = true,
     enableAutoAttach = true,
-    enableAutoAttachmentDetails = true
+    enableAutoAttachmentDetails = true,
+    enableSearchAutoFocus = true
 
 }